Interim checkin of SAFECode port to mainline LLVM and clang.
Applied SAFECode mods to clang.  As a more recent version of 
clang was used, compiling it required an update of LLVM.  This
in turn contained API changes, which necessitated changes to
ExactCheckOpt.cpp.

The resulting version of clang compiles and runs.  The SAFECode 
lit tests have the expected failures due to the removal of 
LoadStoreChecks(.h & .cpp).  Note that some of the SAFECode 
mods to clang are commented out to allow clang to compile in 
the absence of these files.

Expect the following unexpected failure counts in the SAFECode
lit tests:

	core		 10
	cstlib		  1
	formatstrings	  0
	bodiagsuite	805

Note that libLTO is disabled in the safecode/test Makefile.

While all LLVM lit tests pass at present, the lit tests will 
not run on presto unless ulimit -v is increased to 4 GB in 
test/Makefile (this is the llvm directory, not SAFECode).




git-svn-id: https://llvm.org/svn/llvm-project/safecode/trunk@163998 91177308-0d34-0410-b5e6-96231b3b80d8