Fixed PR#11524 and PR#16420. Fixed the local static array bounds checking pass so that it measures the offset from the beginning of the memory object and not from the pointer that is the base of the GEP pointer arithmetic operation.  This fixes PR#16420.
It turns out that proving GEPs safe is more difficult than we thought with
rewrite pointers; one can only eliminate a check if that check and all
operations that generate the pointer argument of the GEP are statically safe.
The GEP check insertion pass incorrectly removed GEP checks without taking
this fact into account.  Disabling this feature fixes PR#11524.

Fixed rewrite-001 to be expected to pass.


git-svn-id: http://llvm.org/svn/llvm-project/safecode/branches/release_32@184646 91177308-0d34-0410-b5e6-96231b3b80d8