Don't force sequential lit testing, use processes.
Not sure what the reason for forcing -j1 is,
but things seem to work fine without it locally.

Only reason I can think of is that on presto python
gets grumpy trying to use 32 threads.  Newer lit
prefers multiprocessing which works around this issue
(and scales better due to GIL), but this was
added after the 3.3 release we're currently using.

Let's see what happens.  If this breaks, I'll
backport the llvm-lit changes from upstream.